Re-organise household incomes and expenses
