--- openfisca_us/tests/policy/baseline/irs/c02900.yaml
@@ -1,18 +1,0 @@
-- name: All inputs used with their haircut rate
-  period: 2020
-  input:
-    c03260: 1
-    care_deduction: 1
-    filer_e03210: 1
-    filer_e03400: 1
-    filer_e03500: 1
-    filer_e00800: 1
-    filer_e03220: 1
-    filer_e03230: 1
-    filer_e03240: 1
-    filer_e03290: 1
-    filer_e03270: 1
-    filer_e03150: 1
-    filer_e03300: 1
-  output:
-    c02900: 8
