+++ openfisca_us/variables/gov/irs/income/posagi.py
@@ -1,0 +1,13 @@
+from openfisca_us.model_api import *
+
+
+class posagi(Variable):
+    value_type = float
+    entity = TaxUnit
+    label = "Positive AGI"
+    unit = USD
+    documentation = "Negative AGI values capped at zero"
+    definition_period = YEAR
+
+    def formula(tax_unit, period, parameters):
+        return max_(tax_unit("adjusted_gross_income", period), 0)
