+++ openfisca_us/variables/household/expense/utilities/count_distinct_utility_expenses.py
