+++ openfisca_us/variables/household/expense/utilities/electricity_expense.py
