+++ openfisca_us/variables/household/income/person/capital_gains/long_term_capital_gains.py
