+++ openfisca_us/variables/household/income/person/capital_gains/short_term_capital_gains.py
