+++ openfisca_us/variables/household/income/person/dividends/dividend_income.py
