+++ openfisca_us/variables/household/income/person/dividends/ordinary_dividend_income.py
