+++ openfisca_us/variables/household/income/person/self_employment/partnership_s_corp_income.py
